+++ core/data/src/main/kotlin/dev/meloda/fast/data/db/AccountsRepository.kt
@@ -6,7 +6,7 @@ interface AccountsRepository {
 
     suspend fun getAccounts(): List<AccountEntity>
 
-    suspend fun getAccountById(userId: Int): AccountEntity?
+    suspend fun getAccountById(userId: Long): AccountEntity?
 
     suspend fun storeAccounts(accounts: List<AccountEntity>)
 }
