+++ core/database/src/main/kotlin/dev/meloda/fast/database/dao/UsersDao.kt
@@ -11,8 +11,8 @@ abstract class UsersDao : EntityDao<VkUserEntity> {
     abstract suspend fun getAll(): List<VkUserEntity>
 
     @Query("SELECT * FROM users WHERE id IN (:ids)")
-    abstract suspend fun getAllByIds(ids: List<Int>): List<VkUserEntity>
+    abstract suspend fun getAllByIds(ids: List<Long>): List<VkUserEntity>
 
     @Query("DELETE FROM users WHERE id IN (:ids)")
-    abstract suspend fun deleteByIds(ids: List<Int>): Int
+    abstract suspend fun deleteByIds(ids: List<Long>): Int
 }
