+++ core/domain/src/main/kotlin/dev/meloda/fast/domain/ConversationsUseCase.kt
@@ -11,9 +11,9 @@ interface ConversationsUseCase {
         offset: Int?,
     ): Flow<State<List<VkConversation>>>
 
-    fun delete(peerId: Int): Flow<State<Int>>
+    fun delete(peerId: Long): Flow<State<Long>>
 
-    fun changePinState(peerId: Int, pin: Boolean): Flow<State<Int>>
+    fun changePinState(peerId: Long, pin: Boolean): Flow<State<Int>>
 
     suspend fun storeConversations(conversations: List<VkConversation>)
 }
