+++ core/model/src/main/kotlin/dev/meloda/fast/model/api/data/VkAttachmentHistoryMessageData.kt
@@ -1,15 +1,15 @@
 package dev.meloda.fast.model.api.data
 
-import dev.meloda.fast.model.api.domain.VkAttachmentHistoryMessage
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import dev.meloda.fast.model.api.domain.VkAttachmentHistoryMessage
 
 @JsonClass(generateAdapter = true)
 data class VkAttachmentHistoryMessageData(
-    @Json(name = "message_id") val messageId: Int,
+    @Json(name = "message_id") val messageId: Long,
     @Json(name = "date") val date: Int,
-    @Json(name = "cmid") val conversationMessageId: Int,
-    @Json(name = "from_id") val fromId: Int,
+    @Json(name = "cmid") val conversationMessageId: Long,
+    @Json(name = "from_id") val fromId: Long,
     @Json(name = "position") val position: Int,
     @Json(name = "attachment") val attachment: VkAttachmentItemData
 ) {
